refactor: move pure helpers out of views/ into utils/
commitMetrics, formatRelativeAge → dates, commitUrl all become
utils/* modules. They don't touch the DOM and shouldn't have been
in views/. Pure file moves + import path updates.
