+++ app/src/coordinator.ts
@@ -23,7 +23,7 @@ import { showRightSidebar, hideRightSidebar } from './views/shell/rightSidebar.j
 import { buildFilePreviewPane, humanLanguageFor } from './views/panes/filePreviewPane.js';
 import { buildCommitPane } from './views/panes/commitPane.js';
 import { buildStreetPane } from './views/panes/streetPane.js';
-import { sameDayCommitCount } from './views/widgets/commitMetrics.js';
+import { sameDayCommitCount } from './utils/commitMetrics.js';
 import { labelFromManifest } from './views/widgets/displayLabel.js';
 import { LIVE_UPDATES } from './config/index.js';
 import { REBUILD_STATUS, LAST_REBUILD_ERROR, LAST_UPDATED_AT } from './store/liveStatus.js';
