+++ app/src/views/panes/commitPane.ts
@@ -17,7 +17,7 @@
 import type { CommitEntry } from '@/types';
 import { makeLucideIcon } from '@/views/widgets/icon.js';
 import { buildPaneHeader } from '@/views/shell/paneHeader.js';
-import { commitUrl } from './commitUrl.js';
+import { commitUrl } from '@/utils/commitUrl.js';
 import { formatRelativeAge, formatFullDate } from '@/utils/dates.js';
 import { fetchCommitDetail } from '@/api/commit.js';
 import { colorForAuthor } from '@/scene/components/fireflies/authorColor.js';
